Welcome to Rump

Rump is ...

L-algebras

Rump.check_LAlgebra — Function

check_LAlgebra(M::Matrix)

Check if the matrix M is a multiplication table of an L-algebra.

Examples

julia> check_LAlgebra([2 2; 1 2])
true

julia> check_LAlgebra([1 2; 1 2])
false

Rump.LAlgebra — Type

LAlgebra(M::Matrix)
LAlgebra(M::Matrix ; check::Bool = false)

The L-algebra with elements {1,…,n} and multiplication table given by M, i.e. i*j = M[i,j].

Check has a default value false. Set check = true to check_LAlgebra(M) before constructing the L-algebra.

See also check_LAlgebra, LAlgebraElem.

Examples

julia> LAlgebra([2 2; 1 2])
LAlgebra([2 2; 1 2])

julia> LAlgebra([2 2; 1 2], check=true)
LAlgebra([2 2; 1 2])

julia> LAlgebra([1 1; 1 2], check=true)
ERROR: [1 1; 1 2] does not define an L-algebra
[...]

Rump.LAlgebraElem — Type

LAlgebraElem(M::Matrix, value::Int)
LAlgebra(M::Matrix ; check::Bool = false)

The element number value in the L-algebra defined by the matrix M. See also LAlgebra.

Examples

julia> A = LAlgebra([2 2; 1 2]);
julia> LAlgebraElem(A, 2)
LAlgebraElem(LAlgebra([2 2; 1 2]), 2)

Base.print — Method

print(x::LAlgebraElem)

When applied to L-algebra elements, prints (using println) the value of x.

Base.:* — Method

*(x::LAlgebraElem, y::LAlgebraElem)

When applied to L-algebra elements of the same L-algebra, multiplies them in the L-algebra.

Examples

julia> A = LAlgebra([2 2; 1 2]);

julia> LAlgebraElem(A,2)*LAlgebraElem(A,1)
LAlgebraElem(LAlgebra([2 2; 1 2]), 1)

julia> LAlgebraElem(A,1)*LAlgebraElem(A,1)*LAlgebraElem(A,1)
LAlgebraElem(LAlgebra([2 2; 1 2]), 1)

julia> LAlgebraElem(A,1)*(LAlgebraElem(A,1)*LAlgebraElem(A,1))
LAlgebraElem(LAlgebra([2 2; 1 2]), 2)

Base.:* — Method

*(S::Union{Set{LAlgebraElem}, Vector{LAlgebraElem}},x::LAlgebraElem)

Return the set of products {s*t:s∈S,t∈T}.

Base.:* — Method

*(S::Union{Set{LAlgebraElem}, Vector{LAlgebraElem}},x::LAlgebraElem)

Return the set of products {s*x∣s∈S}.

order

Base.in — Method

in(x::LAlgebraElem, a::LAlgebra)

Determine wether x is an element of a in the sense that x.algebra==a

julia> A = LAlgebra([2 2; 1 2]);
julia> B = LAlgebra([3 1 3; 1 3 3; 1 2 3]);
julia> LAlgebraElem(A,2) in A
true

julia> LAlgebraElem(B,2) in A
false

Base.size — Method

size(a::LAlgebra)

When applied to an L-algebra returns its cardinality.

Examples

julia> A = LAlgebra([2 2; 1 2]);
julia> size(A)
2

Rump.logical_unit — Function

logical_unit(a::LAlgebra))

Return the logical unit of the L-algebra a, i.e. the element u in a such that xx=xu=1 and 1*x=x for all x in a.

Rump.rand_elem — Function

rand_elem(a::LAlgebra)

Return a random element of a.

Rump.elements — Function

elements(a::LAlgebra)

Return a vector with of all the elements of a.

Examples

julia> a = LAlgebra([2 2; 1 2]);
julia> elements(a)
2-element Vector{LAlgebraElem}:
 LAlgebraElem(LAlgebra([2 2; 1 2]), 1)
 LAlgebraElem(LAlgebra([2 2; 1 2]), 2)