Welcome to Rump

Rump is ...

L-algebras

Rump.check_LAlgebra — Function

check_LAlgebra(M::Matrix)

Check if the matrix M is a multiplication table of an L-algebra.

Examples

julia> check_LAlgebra([2 2; 1 2])
true

julia> check_LAlgebra([1 2; 1 2])
false

Rump.LAlgebra — Type

LAlgebra(M::Matrix)
LAlgebra(M::Matrix ; check::Bool = false)

The L-algebra with elements {1,…,n} and multiplication table given by M, i.e. i*j = M[i,j].

Check has a default value false. Set check = true to check_LAlgebra(M) before constructing the L-algebra.

See also check_LAlgebra, LAlgebraElem.

Examples

julia> LAlgebra([2 2; 1 2])
LAlgebra([2 2; 1 2])

julia> LAlgebra([2 2; 1 2], check=true)
LAlgebra([2 2; 1 2])

julia> LAlgebra([1 1; 1 2], check=true)
ERROR: [1 1; 1 2] does not define an L-algebra
[...]

Rump.LAlgebraElem — Type

LAlgebraElem(M::Matrix, value::Int)
LAlgebra(M::Matrix ; check::Bool = false)

The element number value in the L-algebra defined by the matrix M. See also LAlgebra.

Examples

julia> a = LAlgebra([2 2; 1 2]);
julia> LAlgebraElem(A, 2)
LAlgebraElem(LAlgebra([2 2; 1 2]), 2)

Base.print — Method

print(x::LAlgebraElem)

When applied to L-algebra elements, prints (using println) the value of x.

Base.:* — Method

*(x::LAlgebraElem, y::LAlgebraElem)

When applied to L-algebra elements of the same L-algebra, multiplies them in the L-algebra.

Examples

julia> a = LAlgebra([2 2; 1 2]);
julia> b = LAlgebra([3 1 3; 3 3 3; 1 2 3]);

julia> LAlgebraElem(a,2)*LAlgebraElem(b,2);
ERROR: elements not in the same LAlgebra

julia> LAlgebraElem(a,2)*LAlgebraElem(a,1)
LAlgebraElem(LAlgebra([2 2; 1 2]), 1)

julia> LAlgebraElem(a,1)*LAlgebraElem(a,1)*LAlgebraElem(a,1)
LAlgebraElem(LAlgebra([2 2; 1 2]), 1)

julia> LAlgebraElem(a,1)*(LAlgebraElem(a,1)*LAlgebraElem(a,1))
LAlgebraElem(LAlgebra([2 2; 1 2]), 2)

Base.:* — Method

*(S::Union{Set{LAlgebraElem}, Vector{LAlgebraElem}},x::LAlgebraElem)

Return the set of products {s*t:s∈S,t∈T}.

Base.:* — Method

*(S::Union{Set{LAlgebraElem}, Vector{LAlgebraElem}},x::LAlgebraElem)

Return the set of products {s*x∣s∈S}.

Base.in — Method

in(x::LAlgebraElem, a::LAlgebra)

Determine wether x is an element of a in the sense that x.algebra==a

julia> A = LAlgebra([2 2; 1 2]);
julia> B = LAlgebra([3 1 3; 1 3 3; 1 2 3]);
julia> LAlgebraElem(A,2) in A
true

julia> LAlgebraElem(B,2) in A
false

Base.size — Method

size(a::LAlgebra)

When applied to an L-algebra returns its cardinality.

Examples

julia> A = LAlgebra([2 2; 1 2]);
julia> size(A)
2

Base.length — Method

length(a::LAlgebra)

When applied to an L-algebra returns its cardinality.

Examples

julia> A = LAlgebra([2 2; 1 2]);
julia> length(A)
2

Rump.logical_unit — Function

logical_unit(a::LAlgebra))

Return the logical unit of the L-algebra a, i.e. the element u in a such that xx=xu=1 and 1*x=x for all x in a.

Rump.rand_elem — Function

rand_elem(a::LAlgebra)

Return a random element of a.

Rump.elements — Function

elements(a::LAlgebra)

Return a vector with of all the elements of a.

Examples

julia> a = LAlgebra([2 2; 1 2]);
julia> elements(a)
2-element Vector{LAlgebraElem}:
 LAlgebraElem(LAlgebra([2 2; 1 2]), 1)
 LAlgebraElem(LAlgebra([2 2; 1 2]), 2)

Rump.upset — Function

upset(x::LAlgebraElem)

Given an element x of an L-algebra A, return a vector with of all the elements in A that are bigger than x.

Examples

julia> a = LAlgebra([2 2; 1 2]);
julia> upset(a)
1-element Vector{LAlgebraElem}:
 LAlgebraElem(LAlgebra([2 2; 1 2]), 2)

Rump.downset — Function

downset(x::LAlgebraElem)

Given an element x of an L-algebra A, return a vector with of all the elements in A that are bigger than x.

Examples

julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> downset(x) 2-element Vector{LAlgebraElem}: LAlgebraElem(LAlgebra([2 2; 1 2]), 1) LAlgebraElem(LAlgebra([2 2; 1 2]), 2)

Rump.is_sharp — Function

is_sharp(a::LAlgebra)

Check if A is a sharp L-algebra

Examples

julia-repl julia> A = LAlgebra([2 2; 1 2]); julia> is_sharp(A) true

Rump.is_symmetric — Function

is_symmetric(a::LAlgebra)

Check if A is a symmetric L-algebra

Examples

julia-repl julia> A = LAlgebra([2 2; 1 2]); julia> is_symmetric(A) true

Rump.is_abelian — Function

is_abelian(a::LAlgebra)

Check if A is an abelian L-algebra

Examples

julia-repl julia> A = LAlgebra([2 2; 1 2]); julia> is_abelian(A) false

Rump.is_linear — Function

is_linear(a::LAlgebra)

Check if A is a linear L-algebra

Examples

julia-repl julia> A = LAlgebra([2 2; 1 2]); julia> is_linear(A) false

Rump.is_discrete — Function

is_discrete(a::LAlgebra)

Check if A is a discrete L-algebra

Examples

julia-repl julia> A = LAlgebra([2 2; 1 2]); julia> is_discrete(A) false

Rump.is_semiregular — Function

is_semiregular(a::LAlgebra)

Check if A is a semiregular L-algebra

Examples

julia-repl julia> A = LAlgebra([2 2; 1 2]); julia> is_semiregular(A) true

Rump.is_regular — Function

is_regular(a::LAlgebra)

Check if A is a regular L-algebra

Examples

julia-repl julia> A = LAlgebra([2 2; 1 2]); julia> is_regular(A) true

Rump.is_hilbert — Function

is_hilbert(a::LAlgebra)

Check if A is a hilbert L-algebra

Examples

julia-repl julia> A = LAlgebra([2 2; 1 2]); julia> is_hilbert(A) true

Rump.is_dualBCK — Function

is_dualBCK(a::LAlgebra)

Check if A is a dualBCK-algebra

Examples

julia-repl julia> A = Lalgebra([2 2; 1 2]); julia> is_dualBCK(a) false

Rump.is_KL — Function

is_KL(a::LAlgebra)

Check if a is a KL-algebra

Examples

julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> is_KL(a) true

Rump.is_CL — Function

is_CL(a::LAlgebra)

Check if a is a KL-algebra

Examples

julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> is_CL(a) true

Rump.is_prime_element — Function

is_prime_element(p::LAlgebraElem)

Check if p is a prime element of the L-algebra.

Examples

julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> p = LAlgebraElem(A,1) true julia> n = LAlgebraElem(A,2) false

Rump.prime_elements — Function

prime_elements(a::LAlgebra)

Return a vector with all the prime elements of a

Examples

```julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> prime_elements(a) 1-element Vector{LAlgebraElem}: LAlgebraElem(LAlgebra([2 2; 1 2]), 1)

julia> b = LAlgebra([3 1 3; 3 3 3; 1 2 3]); julia> prime_elements(b) 1-element Vector{LAlgebraElem}: LAlgebraElem(LAlgebra([3 1 3; 3 3 3; 1 2 3]), 1) ```

Rump.is_prime — Function

is_prime(a::LAlgebra)

Check if a is a prime L-algebra

Examples

```julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> is_prime(a) true

julia> b = LAlgebra([3 1 3; 3 3 3; 1 2 3]); julia> is_prime(b) false ```

Rump.is_subLalgebra — Function

is_subLalgebra(s::Union{Set{LAlgebraElem}, Vector{LAlgebraElem}}, a::LAlgebra)

Check if s (either a subset or a vector of eleemtns of a), is a L-subalgebra of a.

Examples

julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> is_subLalgebra([LAlgebraElem(a,1)],a) false julia> is_subLalgebra([LAlgebraElem(a,2)],a) true

Rump.is_invariant — Function

is_invariant(s::Union{Set{LAlgebraElem}, Vector{LAlgebraElem}}, a::LAlgebra)

Check if s (either a subset or a vector of eleemtns of a), is closed under the operation of a.

Examples

julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> is_invariant([LAlgebraElem(a,1)],a) false julia> is_invariant([LAlgebraElem(a,2)],a) true

Rump.is_ideal — Function

is_ideal(s::Union{Set{LAlgebraElem}, Vector{LAlgebraElem}}, a::LAlgebra)

Check if s (either a subset or a vector of eleemtns of a), is an ideal of a.

Examples

julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> is_ideal([LAlgebraElem(a,1)],a) false julia> is_ideal([LAlgebraElem(a,2)],a) true

Rump.subLalgebra_generated_by — Function

subLalgebra_generated_by(s::Union{Set{LAlgebraElem}, Vector{LAlgebraElem}}, a::LAlgebra)

Return the L-subalgebra of a generated by the subset (or the elements in the vector) s.

Examples

```julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> subLalgebrageneratedby([LAlgebraElem(a,1)],a) Set{LAlgebraElem} with 2 elements: LAlgebraElem(LAlgebra([2 2; 1 2]), 1) LAlgebraElem(LAlgebra([2 2; 1 2]), 2)

julia> subLalgebrageneratedby([LAlgebraElem(a,2)],a) Set{LAlgebraElem} with 1 element: LAlgebraElem(LAlgebra([2 2; 1 2]), 2) ```

Rump.ideal_generated_by — Function

ideal_generated_by(s::Union{Set{LAlgebraElem}, Vector{LAlgebraElem}}, a::LAlgebra)

Return the L-subalgebra of a generated by the subset (or the elements in the vector) s.

Examples

```julia-repl julia> a = LAlgebra([2 2; 1 2]); julia> idealgeneratedby([LAlgebraElem(a,1)],a) Set{LAlgebraElem} with 2 elements: LAlgebraElem(LAlgebra([2 2; 1 2]), 1) LAlgebraElem(LAlgebra([2 2; 1 2]), 2)

julia> idealgeneratedby([LAlgebraElem(a,2)],a) Set{LAlgebraElem} with 1 element: LAlgebraElem(LAlgebra([2 2; 1 2]), 2) ```